pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "emergency-mesh-router"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "openenv-core",
    "numpy"
]